requirements.txt
asgiref==3.10.0
channels==4.1.0
channels-redis==4.2.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.11
inflection==0.5.1
modeltranslation==0.25
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg2==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==5.2.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
twilio==9.3.0
